Refactor size retrieval methods to use Length() instead of Size()
- Updated various classes in the Visualization module to replace calls to Size() with Length() for consistency and clarity.
- Adjusted the handling of texture sets, selection managers, and other data structures to ensure proper size retrieval.
- Added viewport offset methods in V3d_View for framebuffer rendering.
- Included additional checks for finite values in SelectMgr_Frustum to enhance robustness.

+++ src/DataExchange/TKXSBase/Transfer/Transfer_ProcessForFinder_0.cxx
@@ -141,8 +141,8 @@ void Transfer_ProcessForFinder::Clean()
 
 void Transfer_ProcessForFinder::Resize(const int nb)
 {
-  if (nb > themap.NbBuckets())
-    themap.ReSize(nb);
+  if (static_cast<size_t>(nb) > themap.NbBuckets())
+    themap.ReSize(static_cast<size_t>(nb));
 }
 
 //=================================================================================================
